Extract completion logic outside of editor
